Remove babel-preset-react-optimize
The babel-preset-react-optimize preset doesn't reduce the size of our bundles. Documentation seems to indicate that it will only alter the output of your transpiled React components, so I'm not seeing the value added. I'm proposing we remove it as the first of a few improvements to this preset. If rejected, I can at least learn why we had it in the first place.